# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document API",
    "version": "1.0.0",
    "description": "API documentation for document-related operations"
  },
  "tags": [
    {
      "name": "Document",
      "description": "Operations related to documents"
    },
    {
      "name": "Auth",
      "description": "Operations related to user authentication"
    }
  ],
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token for authentication"
      }
    },
    "schemas": {
      "Version": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "required": true
          },
          "modifiedBy": {
            "type": "string",
            "required": true
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time",
            "default": "Date.now"
          }
        }
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "required": true
          },
          "content": {
            "type": "string",
            "required": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "default": "null"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "file": {
            "type": "object",
            "properties": {
              "data": {
                "type": "string",
                "format": "binary",
                "required": true
              },
              "contentType": {
                "type": "string",
                "required": true
              },
              "originalName": {
                "type": "string",
                "required": true
              }
            }
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Version"
            }
          }
        },
        "required": ["id", "content", "file"]
      },
      "PaginatedDocumentList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Document"
            }
          },
          "info": {
            "type": "object",
            "properties": {
              "totalItems": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              },
              "currentPage": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              }
            }
          }
        }
      },
      "LoginRequestBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "required": true
          },
          "password": {
            "type": "string",
            "required": true
          }
        },
        "required": ["email", "password"]
      },
      "RegisterRequestBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "required": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "required": true
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "required": true
          }
        },
        "required": ["name", "email", "password"]
      },
      "UserSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "user"]
          }
        },
        "required": ["id", "name", "email", "role"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "$ref": "#/components/schemas/UserSchema",
            "description": "User information"
          },
          "accessToken": {
            "type": "string",
            "description": "Access token for authentication"
          }
        },
        "required": ["user", "accessToken"]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "$ref": "#/components/schemas/UserSchema",
            "description": "User information"
          }
        },
        "required": ["user"]
      }
    }
  },
  "paths": {
    "/documents": {
      "get": {
        "summary": "Get documents",
        "tags": ["Document"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string",
              "minimum": 1
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string",
              "minimum": 1
            },
            "description": "Items per page"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search term"
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["DESC", "ASC"]
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDocumentList"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/documents/upload": {
      "post": {
        "summary": "Upload a document",
        "tags": ["Document"],
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "file",
            "in": "formData",
            "schema": {
              "type": "string",
              "format": "binary"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/documents/{documentId}": {
      "put": {
        "summary": "Update a document",
        "tags": ["Document"],
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "name": "file",
            "in": "formData",
            "schema": {
              "type": "string",
              "format": "binary"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "summary": "Soft delete a document",
        "tags": ["Document"],
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User login",
        "tags": ["Auth"],
        "consumes": ["application/json"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "User registration",
        "tags": ["Auth"],
        "consumes": ["application/json"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  }
}